Set parent of mortar segment elements
This prevents failed assertions in libMesh in debug mode
during MeshBase::prepare_for_use when refinement is present.
These assertions occur when the level of an element and its
interior_parent don't match. When an element doesn't have a parent
its level is always considered zero. However, here if I set
the parent of the mortar segment equal to the parent of the slave
element from which it is initially spawned, then the level of
the mortar segment and its initerior_parent are guaranteed to match.

There is a problem with this. If prepare_for_use is called with
find_neighbors=true, then there is a failed assertion when the mesh
tries to determine which child the mortar segment is of the parent
slave_elem. Obviously it is not its child. My solution for now
is to simply called prepare_for_use with find_neighbors=false. At
some point I should come up with a solution that is conceptually
sound and consequently more robust. That solution may be, as pointed
out by @roystgnr in libMesh/libmesh#2289 to get rid of some of the
libMesh assertions, perhaps the assertion that an element and
its interior_parent should have matching levels