Make getFieldVar go through getVarHelper
This ensures that all the smart code that's been added to
getVarHelper gets applied when getFieldVar is called.
Specifically for my immediate application, this checks
the map from new to deprecated coupled vars which ensures
that a variable is retrieved when the deprecated coupled
variable name is used. Coupleable::coupled uses the getFieldVar
code path which is what was biting me in the butt for INS
based simulations