Create and define the FunctorInterface
GenericFunctors inherit from FunctorInterface and these
are the building block for constant functors and material
property functors. Variables also inherit from FunctorInterface.

This commit also adds the FaceInfo operator() overload to
the FunctorInterface. This overload takes FaceInfo, Limiter,
and bool element_is_upwind arguments as a tuple to allow
unambiguous creation of face values. Corresponding to this
addition we have created a new getInternalFaceValue overload
in MooseVariableFV that uses cell values, gradient, and the limiter
to compute the face value