Remove PWCNSFVStrongBC
In brief trials, I have not been able to get the energy equation
to converge at the correct rate. It's actually fully possible that I
simply had not refined the mesh enough yet. But anyway this current
feature branch is not about Euler/Navier-Stokes; it's about introducing
functor properties and solving the mass and momentum equations in a
weakly compressible context is enough of a physics demonstration of
the usefulness of functor properties. Incorporating a weakly compressible
energy equation can be a part of a feature branch more focused
on the `navier_stokes` module