Add time argument to functors
For now I chose to go with an unsigned int for the time argument.
It corresponds to the solution state. This can be used to query
for old and older variable values and can also be easily used
to determine Real time values by querying `FEProblemBase` (e.g.
`time()` or `timeOld()` methods)