Make subdomain ID the first tuple arg for every space arg
I started doing this because I thought I wanted a user lambda in
a material to be able to query the space argument in a consistent
way for the subdomain ID. However, I think that if a user is querying
the space argument for data, I've done something wrong in my design.
A space argument should only be forwarded in a user functor material
lambda and then if so, the developer implementer of the callee should
know how to process the space argument in an appropriate way.