Overhaul FaceArg functor overload
A comment that Guillaume made really made me think about the design of
this overload. It needed to be revamped to really take sidedness into
account. So now the SubdomainID argument has been changed to be a
std::pair<SubdomainID, SubdomainID>. This qualitative change now makes
it conceptually sound in my mind to call the ElemFromFaceArg overload in
FunctorMaterialProperty when the FaceArg overload is called. And then we
can do a macroscopic interpolation between the results which is an
ability that others clamored for in the beginning