Fix terrible reference AttribVar object
You should not create a reference Attribute object that is already a
match of real life scenarios. Because when you clone the reference
condition you will copy construct that realistic value. And then in this
case if you have a MooseObject that is not a MooseVariableInterface,
then the clone will not have its _val init'd from it, leaving it in the
realistic state which can be matched during the course of the
simulation. For instance if I query with a variable number of 0 I'll get
a hit on this AttribVar object that was copy constructed to have a value
of 0. I love debugging this for hours