Add free-slip friction PINSFV test
This test almost perfectly mimics the 'problematic' input that Joe sent
us. It has friction, a symmetry boundary, and a free slip wall boundary.
The velocity field should be constant throughout the entire domain, and
that is what we see when we have two term boundary expansion turned on
which allows the pressure gradient to be constant and to perfectly
balance the friction term in the momentum balance