Remove RZ integration from b2 computation
This still keeps optimal convergence for the new slate of tests
but also fails on the RZ RC test when xmin is set to 0. I'll
reiterate that that MMS test passes when I switch to standard
body forces. Maybe the RZ integration of area and volume doesn't
matter, e.g. the bug is elsewhere