Use functor material property porosity for reconstruction
This required some refactoring to allow passing a `consumer` parameter
to `Moose::FV::reconstruct` such that we can indicate the proper block
restriction we want for the `input_functor`. I also added an error
if a user attempts to reconstruct porosity from a variable.

The reconstruct test now passes in parallel, hoorah. Not brave enough
to try distributed yet