Include Newmark-beta in the dynamic mortar contact constraints (#19671)
The stabilization of the normal contact constraint requires adding the
time discretization integration scheme in the constraint. This need is
obvious when running a challenging problem such as the ironing problem
when large deformation and nodal element passage come into play.

For now, we only 'allow' Newmark-beta integrator, which is fairly
standard for solid mechanics. The alpha parameter of HHT can be added
in the future.