Add missing mesh advection term
This term comes because to convert to the ALE system you have this:

partial phi/partial t -> partial phi/partial t - mesh_vel * grad phi

The last term is *not* the same as -\nablda \cdot (mesh_vel * phi).
So what we end up implementing is:

-mesh_vel * grad_phi -> -(\nabla\cdot(mesh_vel*phi) - phi\nabla\cdot(mesh_vel))

The first term is covered by the flux advection kernel. We now
introduce an elemental kernel to cover the second term